+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.3.0] - 2024-01-10
+
+### Changed
+
+- Added new MythicRPC function for searching a callbacks' edges
+- Added new MythicRPC function for created a task in a specific callback
+- Added new Payload definition function for `onNewCallback`
+- Updated Container version to 1.2.0
+
 ## [v1.2.1] - 2023-12-05
 
 ### Changed